chore(consolidate): problema and tests
Consolidate questions and tests for Stack's minimum in contant time.
Sum tree checker.
MaxNumber with all digits in a number
