+++ InterviewQuestions.sln
@@ -31,15 +31,9 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "QueueTaleOfTwoStacks", "Que
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TheMaximumNumberWithAllDigitsInANumber", "TheMaximumNumberWithAllDigitsInANumber\TheMaximumNumberWithAllDigitsInANumber.csproj", "{C3C2F4DE-E3B6-4CE7-9479-31B90FFD8FAF}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TheMaximumNumberWithAllDigitsInANumberTester", "TheMaximumNumberWithAllDigitsInANumberTester\TheMaximumNumberWithAllDigitsInANumberTester.csproj", "{938F4D8C-9101-4965-8D63-05018E3A5743}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SumTreeQuestion", "SumTree\SumTreeQuestion.csproj", "{C89EB85F-83F9-435A-8914-3E3B912E2F53}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SumTreeQuestionTester", "SumTreeQuestionTester\SumTreeQuestionTester.csproj", "{F4ACF20C-CD60-4820-9507-BCF53496E22B}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "StackMinimumQuestion", "StackMinimumQuestion\StackMinimumQuestion.csproj", "{3C5AD6E6-B303-459A-867C-59D439B7DE76}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "StackMinimumQuestionTester", "StackMinimumQuestionTester\StackMinimumQuestionTester.csproj", "{6B3B757F-DEA0-457C-91AB-3743AECC25F2}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{27CB879C-6B9E-45B2-9333-1630FD98F79C}"
 	ProjectSection(SolutionItems) = preProject
@@ -129,26 +123,14 @@ Global
 		{C3C2F4DE-E3B6-4CE7-9479-31B90FFD8FAF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C3C2F4DE-E3B6-4CE7-9479-31B90FFD8FAF}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C3C2F4DE-E3B6-4CE7-9479-31B90FFD8FAF}.Release|Any CPU.Build.0 = Release|Any CPU
-		{938F4D8C-9101-4965-8D63-05018E3A5743}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{938F4D8C-9101-4965-8D63-05018E3A5743}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{938F4D8C-9101-4965-8D63-05018E3A5743}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{938F4D8C-9101-4965-8D63-05018E3A5743}.Release|Any CPU.Build.0 = Release|Any CPU
 		{C89EB85F-83F9-435A-8914-3E3B912E2F53}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{C89EB85F-83F9-435A-8914-3E3B912E2F53}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C89EB85F-83F9-435A-8914-3E3B912E2F53}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C89EB85F-83F9-435A-8914-3E3B912E2F53}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F4ACF20C-CD60-4820-9507-BCF53496E22B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F4ACF20C-CD60-4820-9507-BCF53496E22B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F4ACF20C-CD60-4820-9507-BCF53496E22B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F4ACF20C-CD60-4820-9507-BCF53496E22B}.Release|Any CPU.Build.0 = Release|Any CPU
 		{3C5AD6E6-B303-459A-867C-59D439B7DE76}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{3C5AD6E6-B303-459A-867C-59D439B7DE76}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3C5AD6E6-B303-459A-867C-59D439B7DE76}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3C5AD6E6-B303-459A-867C-59D439B7DE76}.Release|Any CPU.Build.0 = Release|Any CPU
-		{6B3B757F-DEA0-457C-91AB-3743AECC25F2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{6B3B757F-DEA0-457C-91AB-3743AECC25F2}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6B3B757F-DEA0-457C-91AB-3743AECC25F2}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{6B3B757F-DEA0-457C-91AB-3743AECC25F2}.Release|Any CPU.Build.0 = Release|Any CPU
 		{C788EAE6-D9DD-473F-A551-62DFD2EE2689}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{C788EAE6-D9DD-473F-A551-62DFD2EE2689}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C788EAE6-D9DD-473F-A551-62DFD2EE2689}.Release|Any CPU.ActiveCfg = Release|Any CPU
