+++ StackMinimumQuestion/packages.config
