+++ TheMaximumNumberWithAllDigitsInANumber/packages.config
